# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  models.UpdateUser:
    properties:
      bio:
        type: string
      imageUrl:
        type: string
      name:
        type: string
    type: object
  models.UserIdsRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
  models.UserModel:
    properties:
      _id:
        type: string
      bio:
        type: string
      email:
        type: string
      followers:
        items:
          type: string
        type: array
      following:
        items:
          type: string
        type: array
      imageUrl:
        type: string
      name:
        type: string
    required:
    - email
    type: object
host: localhost:5000
info:
  contact: {}
  description: this is  Swagger docs for Social chat API created with golang and fiber
  title: Fiber Golang Rest API
  version: "1.0"
paths:
...
  /user/{id}/following:
    patch:
      consumes:
      - application/json
      description: follow or un follow a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Follow/UnFollow User
      tags:
      - Users
  /user/Update/{id}:
    patch:
      consumes:
      - application/json
      description: update user details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: 'details '
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: update user data
      tags:
      - Users
  /user/batch:
    post:
      consumes:
      - application/json
      description: Get a list of users' public info by an array of user IDs
      parameters:
      - description: list of user ids
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/models.UserIdsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get multiple users by their IDs
      tags:
      - Users
  /user/delete/{id}:
    delete:
      consumes:
      - application/json
      description: delete user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: delete user
      tags:
      - Users
  /user/getUser/{id}:
    get:
      consumes:
      - application/json
      description: GetUser Details By ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get User By ID
      tags:
      - Users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"